bookstore/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
sqlalchemy==2.0.31
PyJWT==2.4.0
passlib==1.7.4
bcrypt==4.1.3
sqlmodel==0.0.19
pytest-cov